muon: Newton-Schulz runs in the momentum dtype
The NS input was upcast to fp32 at both call sites (engine kernel and
its reference twin) since the original port, with no recorded
rationale — on sm90 cuBLAS serves that with sm80 fp32 compat GEMMs
and no tensor cores: 225 ms/step of optimizer NS time on the traced
llama3_8b muon cell. Momentum math was already bf16 by design; the
orthogonalization now follows it on both sides together, so
engine-twin parity is preserved at bf16. Full certification battery
green with zero golden re-blesses (muon properties, engine-vs-
reference, family goldens, engine parity).

+++ src/dataflow_training/kernels/muon.py
@@ -51,7 +51,7 @@ def _muon_step_aten(kctx, w, g, m, *, shape, lr, beta, eps,
     m.mul_(beta).add_(gm)
     eff = gm.add(m, alpha=beta)             # nesterov
     eff3 = eff.view(shape if len(shape) == 3 else (1, *shape))
-    o = ns_orthogonalize_batched(eff3.float(), eps=eps)
+    o = ns_orthogonalize_batched(eff3, eps=eps)
     if weight_decay:
         w32 = w.float().mul_(1.0 - lr * weight_decay)
         w.copy_(w32.to(w.dtype))
